curahealthcare/requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
blessed==1.20.0
boto3==1.34.88
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.3
crispy-bootstrap4==2024.1
Django==5.0.1
django-crispy-forms==2.1
django-dotenv==1.4.2
django-sslserver==0.22
django-storages==1.14.2
django-widget-tweaks==1.5.0
idna==3.7
jmespath==1.0.1
pathspec==0.10.1
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.9.0
urllib3==1.26.18
wcwidth==0.1.9
